rest api with angular

+++ backend/controllers/productController.js
@@ -14,38 +14,39 @@ const getProduct = async (req, res) => {
 }
 
 const createProduct = async (req, res) => {
-  const { name, description, price, inStock } = req.body.product
+  const { name, description, price, inStock } = JSON.parse(req.body.product)
   const product = await ProductModel.create({ 
     name, 
     description, 
     price, 
     inStock,
-    imageUrl: req.file ? `${req.protocol}://${req.get('host')}/public/images/${req.file.filename}` : ''
+    imageUrl: req.file ? `${req.protocol}://${req.get('host')}/images/${req.file.filename}` : ''
   })
 
-  res.json({ product })
+  res.json(product)
 }
 
 const updateProduct = async (req, res) => {
-  const { name, description, price, inStock } = req.body
-  await ProductModel.findByIdAndUpdate(req.params.id, { 
-    name, 
-    description, 
-    price, 
-    inStock,
-    imageUrl: req.file ? `${req.protocol}://${req.get('host')}/public/images/${req.file.filename}` : ''
-  })
+  const product = req.file ? {
+    ...JSON.parse(req.body.product),
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  } : { ...req.body }
+
+  await ProductModel.findByIdAndUpdate(req.params.id, product)
 
   res.json({ message: 'Modified!' })
 }
 
 const deleteProduct = async (req, res) => {
   const productToDelete = await ProductModel.findById(req.params.id)
-  const filename = productToDelete.imageUrl.split('/public/images/')[1]
-  fs.unlink(`public/images/${filename}`, async () => {
-    await ProductModel.deleteOne(req.params.id)
-    res.json({ message: 'Deleted!' })
-  })
+
+  if(productToDelete) {
+    const filename = productToDelete.imageUrl.split('/images/')[1]
+    fs.unlink(`public/images/${filename}`, async () => {
+      await ProductModel.deleteOne({_id: req.params.id})
+      res.json({ message: 'Deleted!' })
+    })
+  }
 }
 
 module.exports = {
